feat(nextjs): withHoneybadger wrapper for route handlers and middleware
